backend refactoring completed with simultaneous listener for socker and metrics port
