+++ vm/server.ts
@@ -1,65 +1,22 @@
-import express from 'express';
-import { Express, Request, Response, NextFunction } from 'express';
-import bodyParser from 'body-parser';
 import fs from 'fs';
-import api from './server/routes/api';
-import http from 'http';
+import metricsServer from './server/metricsServer';
+import socketServer from './server/socketServer';
+import { SOCKET_PATH } from './server/constants';
+import { METRICS_PORT } from './server/constants';
 
-// Unix socket
-const socket = '/run/guest-services/backend.sock';
-
-// After a server is done with the unix domain socket, it is not automatically destroyed.
-// You must instead unlink the socket in order to reuse that address/path.
-// To do this, we delete the file with fs.unlinkSync()
+// If a socket file exists, delete it for reusability.
 try {
-  if (fs.existsSync(socket)) {
-    console.log('UNIX socket file exists.');
-    fs.unlinkSync(socket);
-    console.log('Deleted the UNIX socket file.');
+  if (fs.existsSync(SOCKET_PATH)) {
+    fs.unlinkSync(SOCKET_PATH);
   }
 } catch (err) {
-  console.error('An error occurred while deleting the socket file:', err);
+  console.error('Error deleting socket file:', err);
 }
 
-const app: Express = express();
-
-app.use(bodyParser.json());
-
-app.get('/hello', (req, res) => {
-  res.send(`🚀 Server listening on ${socket}`);
+metricsServer.listen(METRICS_PORT, () => {
+  console.log(`🚀 Server listening on Port ${METRICS_PORT}`);
 });
 
-const server = http.createServer(app);
-server.listen(socket, () => {
-  console.log(`🚀 Server listening on ${socket}`);
+socketServer.listen(SOCKET_PATH, () => {
+  console.log(`🚀 Server listening on Socket ${SOCKET_PATH}`);
 });
-
-app.use('/api', api);
-
-// catch-all route handler
-app.use((_req: Request, res: Response): unknown =>
-  res.status(404).send("This is not the page you're looking for...")
-);
-
-// global error handler
-app.use(
-  (
-    err: unknown,
-    _req: Request,
-    res: Response,
-    _next: NextFunction
-  ): unknown => {
-    const defaultErr = {
-      log: `Express error handler caught unknown middleware error ${err}`,
-      status: 500,
-      message: { err: 'An error occurred' },
-    };
-    const errorObj = Object.assign({}, defaultErr, err);
-    console.log(errorObj.log);
-    return res.status(errorObj.status).json(errorObj.message);
-  }
-);
-
-// app.listen(3333, () => {
-//   console.log('listening on Port 3333...');
-// });
