+++ vm/server/constants.ts
@@ -1,0 +1,2 @@
+export const SOCKET_PATH = '/run/guest-services/backend.sock';
+export const METRICS_PORT = 3333;
